eff_tunnel: fix tysyn, raise non-term indices
There was an issue where indices were too high. It was caused by something like
    |- /\.\:[T]_0\:[T]_1.0 : [/\.\[T]_0.[T]_1]_[]
where it should be
    |- /\.\:[T]_0\:[T]_1.0 : [/\.\[T]_0.[T]_0]_[]
The reason being that fn's can introduce a new variable in terms, but not
in types. So any references to variables outside the function body will
need to be raised by one when synthesizing the type.

So in the example above, we have
  a,T,T |- 0 : [T]_[]
  a,T |- \:[T]_1.0 : [\:[T]_0.T]_[]
  a |- \:[T]_0.\:[T]_1.0 : [\[T]_0.\[T]_0.T]_[]
  |- /\.\:[T]_0.\:[T]_1.0 : [/\.\[T]_0.\[T]_0.T]_[]